Update for CD support.
This provides support for loceee's variant of cocoaDialog which can run over the OS X loginwindow. FauxLogout functionality has not been removed but all the applescript has been replaced with equivalent cocoaDialog code and "should be" functionally equivalent.

WARNING: I do NOT have a JSS and thus this fork is HIGHLY UNTESTED and it probably riddled full of bugs.